reconstruct review page style
