+++ app/src/main/AndroidManifest.xml
@@ -23,9 +23,6 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".activity.PinlunActivity"
-            android:exported="false" />
-        <activity
             android:name=".activity.FindDetailActivity"
             android:launchMode="singleTask"
             android:exported="false" />
